+++ shortcut_composer/api_krita/wrappers/document.py
@@ -3,6 +3,7 @@
 
 from dataclasses import dataclass
 from typing import List, Protocol
+from ..enums import NodeType
 
 from .node import Node, KritaNode
 
@@ -12,6 +13,7 @@ class KritaDocument(Protocol):
 
     def activeNode(self) -> KritaNode: ...
     def setActiveNode(self, node: KritaNode): ...
+    def createNode(self, name: str, node_type: NodeType) -> KritaNode: ...
     def topLevelNodes(self) -> List[KritaNode]: ...
     def resolution(self) -> int: ...
     def currentTime(self) -> int: ...
@@ -34,6 +36,21 @@ class Document:
     def active_node(self, node: Node) -> None:
         """Set active `Node`."""
         self.document.setActiveNode(node.node)
+
+    def create_node(self, name: str, node_type: NodeType) -> Node:
+        """
+        Create a Node.
+        
+        IMPORTANT: Created node must be then added to node tree to be usable from Krita.
+        For example with add_child_node() method of Node Class.
+
+        When relevant, the new Node will have the colorspace of the image by default;
+        that can be changed with Node::setColorSpace.
+
+        The settings and selections for relevant layer and mask types can also be set
+        after the Node has been created.
+        """
+        return Node(self.document.createNode(name, node_type.value))
 
     @property
     def current_time(self) -> int: